Move filter text input to separate component
